Add files via upload
This one is for beginners who are eager to learn What Gans is?? How to implement Gans in Tensorflow using Keras, so I have upload slides for Gans explanation and 2 files for training Gans. In one model, I have trained it using CIFAR 10  and Fashion-Mnist dataset. These models will be building blocks for learners to dive in Gans and explore more to generate Deep faces and much more interesting applications.  Here is the link of my Youtube Channel video in which I have explained Gans and its Implementation in Tensorflow using Keras.